requirements.txt
ipykernel
yfinance
pandas
Ta-Lib
scikit-learn
numpy
jupyter
plotly
nbformat
ipywidgets
dash
fear-and-greed-crypto
pycoingecko
matplotlib
nbstripout
mlflow
tabulate
shap
wheel
setuptools
quantreo
quantstats
Boruta
coinmetrics-api-client